refactor: extract state helpers
